third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGCircleElement-dom-cy-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'cy' attribute of the SVGCircleElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS circleElement.getAttribute('cy') is "-50"
PASS circleElement.getAttribute('cy') is "150"
PASS successfullyParsed is true

TEST COMPLETE
